LUN-58: Add sharing option for results
