+++ release-notes/unreleased/LUN-58-sharing-option-for-results.yml
@@ -1,0 +1,6 @@
+issue_key: LUN-58
+show_in_stores: true
+platforms:
+  - android
+  - ios
+de: Möglichkeit zum Teilen von Ergebnissen hinzugefügt
