+++ src/routes/exercise-finished/ExerciseFinishedScreen.tsx
@@ -1,15 +1,15 @@
 import { RouteProp } from '@react-navigation/native'
 import { StackNavigationProp } from '@react-navigation/stack'
 import React, { ReactElement } from 'react'
-import { StatusBar } from 'react-native'
 import { heightPercentageToDP as hp } from 'react-native-responsive-screen'
 import styled from 'styled-components/native'
 
-import { CheckCircleIconWhite, ListIcon, RepeatIcon } from '../../assets/images'
-import Button from '../components/Button'
-import { BUTTONS_THEME, ExerciseKeys, EXERCISES } from '../constants/data'
-import labels from '../constants/labels.json'
-import { RoutesParams } from '../navigation/NavigationTypes'
+import { CheckCircleIconWhite, ListIcon, RepeatIcon } from '../../../assets/images'
+import Button from '../../components/Button'
+import { BUTTONS_THEME, ExerciseKeys, EXERCISES } from '../../constants/data'
+import labels from '../../constants/labels.json'
+import { RoutesParams } from '../../navigation/NavigationTypes'
+import ShareSection from './components/ShareSection'
 
 const Root = styled.View`
   background-color: ${prop => prop.theme.colors.background};
@@ -18,7 +18,7 @@ const Root = styled.View`
 `
 const UpperSection = styled.View`
   width: 140%;
-  height: 60%;
+  height: 45%;
   background-color: ${prop => prop.theme.colors.primary};
   border-bottom-left-radius: ${hp('60%')}px;
   border-bottom-right-radius: ${hp('60%')}px;
@@ -43,7 +43,7 @@ interface InitialSummaryScreenProps {
   navigation: StackNavigationProp<RoutesParams, 'InitialSummary'>
 }
 
-const InitialSummaryScreen = ({ navigation, route }: InitialSummaryScreenProps): ReactElement => {
+const ExerciseFinishedScreen = ({ navigation, route }: InitialSummaryScreenProps): ReactElement => {
   const { exercise, discipline, results } = route.params.result
   const [message, setMessage] = React.useState<string>('')
 
@@ -79,8 +79,6 @@ const InitialSummaryScreen = ({ navigation, route }: InitialSummaryScreenProps):
 
   return (
     <Root>
-      <StatusBar barStyle='light-content' />
-
       <UpperSection>
         <CheckCircleIconWhite />
         <MessageContainer>
@@ -100,8 +98,9 @@ const InitialSummaryScreen = ({ navigation, route }: InitialSummaryScreenProps):
         buttonTheme={BUTTONS_THEME.outlined}
         onPress={repeatExercise}
       />
+      <ShareSection discipline={discipline} results={results} />
     </Root>
   )
 }
 
-export default InitialSummaryScreen
+export default ExerciseFinishedScreen
